End of task 1 and 2
 mixer + e_NTU correlation in toolbox
